PanTilt helper functions and simple training test
Unfortunately the ultrasonic sensor only has about a 25 degrees (either
side) useful-reading-range, so I've been unable to calculate distances
with it's data. The current WIP training just stops the robot from
falling off an edge it's right next to. But it's a start! I've also
ordered a cheap laser range finder, which could potentially be used
instead / as well.